cambios pep2 funcional

+++ PEP2/src/practica/ui/EvolutionListener.java
@@ -2,8 +2,8 @@ package practica.ui;
 
 public interface EvolutionListener {
     void onGeneration(int generation,
-                      int bestOfGeneration,
-                      int globalBestSoFar,
+                      double bestOfGeneration,
+                      double globalBestSoFar,
                       double averageFitness,
                       Object bestChromosomeOfGeneration);
 }
