+++ PEP2/src/practica/ui/GAResult.java
@@ -2,11 +2,11 @@ package practica.ui;
 
 public class GAResult {
     private final Object best;
-    private final int[] bestOfGeneration;
-    private final int[] globalBestSoFar;
+    private final double[] bestOfGeneration;
+    private final double[] globalBestSoFar;
     private final double[] averageFitness;
 
-    public GAResult(Object best, int[] bestOfGeneration, int[] globalBestSoFar, double[] averageFitness) {
+    public GAResult(Object best, double[] bestOfGeneration, double[] globalBestSoFar, double[] averageFitness) {
         this.best = best;
         this.bestOfGeneration = bestOfGeneration;
         this.globalBestSoFar = globalBestSoFar;
@@ -14,7 +14,7 @@ public class GAResult {
     }
 
     public Object getBest() { return best; }
-    public int[] getBestOfGeneration() { return bestOfGeneration; }
-    public int[] getGlobalBestSoFar() { return globalBestSoFar; }
+    public double[] getBestOfGeneration() { return bestOfGeneration; }
+    public double[] getGlobalBestSoFar() { return globalBestSoFar; }
     public double[] getAverageFitness() { return averageFitness; }
 }
